requirements.txt
anthropic
moviepy
openai-whisper @ git+https://github.com/openai/whisper.git
python-dotenv
setuptools-rust
streamlit